moved motor setup to chassis ctrl, added skeleton for drive mode in receiver main, left todo notes for drive mode on receiver side as well as toggling drive & range finder mode on controller side

+++ scout_bot_controller/src/main.cpp
@@ -45,11 +45,13 @@ void loop() {
     uint8_t *left_joy_y_payload = (uint8_t *) &left_joy_y;
     receiver_service->getCharacteristic(LEFT_JS_Y_CHARACTERISTIC_UUID)->writeValue(left_joy_y_payload, sizeof(float));
     
-    uint32_t range_enabled = button_read_sel(&buttons, 0);
+    /* todo; this needs to be fixed, make this a toggle instead, rather than sending across current button state (pressed or not pressed). toggle & transmit range finder enabled mode */
+    uint8_t range_enabled = button_read_sel(&buttons, 0);
     uint8_t *range_enabled_payload = (uint8_t *) &range_enabled;
-    receiver_service->getCharacteristic(RANGE_ENABLE_CHARACTERISTIC_UUID)->writeValue(range_enabled_payload, sizeof(uint32_t));
+    receiver_service->getCharacteristic(RANGE_ENABLE_CHARACTERISTIC_UUID)->writeValue(range_enabled_payload, sizeof(uint8_t));
 
-    uint32_t drive_mode = button_read_sel(&buttons, 1);
+    /* todo; this needs to be fixed, make this a toggle instead, rather than sending across current button state (pressed or not pressed). toggle & transmit drive mode state */
+    uint8_t drive_mode = button_read_sel(&buttons, 1);
     uint8_t *drive_mode_payload = (uint8_t *) &drive_mode;
-    receiver_service->getCharacteristic(DRIVE_MODE_CHARACTERISTIC_UUID)->writeValue(drive_mode_payload, sizeof(uint32_t));
+    receiver_service->getCharacteristic(DRIVE_MODE_CHARACTERISTIC_UUID)->writeValue(drive_mode_payload, sizeof(uint8_t));
 }
